chatView done, maybe fix the date, and Important fix theuserINput
